# openapi.yaml
openapi: 3.1.0
info:
  title: ProjectIntelligence AU API
  version: 1.0.0
  description: |
    API for ProjectIntelligence AU (PI-AU) — a SaaS product delivering weekly curated digests
    of roofing-relevant Development Applications to Sydney subcontractor businesses.

    **Wedge:** The Sunday-night roofing DA digest for Sydney subbies — 15 LGAs, 5–15 leads, AUD 199/mo, signup in 60 seconds.

    **Tech Stack Contract:** docs/00-tech-stack.md (LOCKED 2026-04-28)

    **Scale Tier:** preview
  contact:
    name: ProjectIntelligence AU Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://pi-au.example.com/api
    description: Production (preview)
...
paths:
...
  /account/me:
    get:
      summary: Get current account profile
      operationId: accountGetMe
      tags:
        - Account
      description: |
        Retrieve the authenticated user's account profile.
        Requires active Lucia session.
      security:
        - LuciaSession: []
      responses:
        '200':
          description: Account profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update account profile
      operationId: accountUpdateMe
      tags:
        - Account
      description: |
        Update the authenticated user's account profile.
        Requires active Lucia session.
      security:
        - LuciaSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /account/lga-bundles:
    get:
      summary: Get user's selected LGA bundles
      operationId: accountGetLgaBundles
      tags:
        - Account
      description: |
        Retrieve the authenticated user's selected LGA bundle IDs.
        **FR-020 (LGA selection)**: Enable self-serve LGA bundle customization.
      security:
        - LuciaSession: []
      responses:
        '200':
          description: List of selected LGA bundle IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LgaBundlesResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update user's selected LGA bundles
      operationId: accountUpdateLgaBundles
      tags:
        - Account
      description: |
        Update the authenticated user's selected LGA bundles.
        Accepts an array of bundle IDs.
      security:
        - LuciaSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLgaBundlesRequest'
      responses:
        '200':
          description: LGA bundles updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /account/saved-query:
    get:
      summary: Get user's saved search query
      operationId: accountGetSavedQuery
      tags:
        - Account
      description: |
        Retrieve the authenticated user's saved search query text.
        The embedding is stored server-side and used in the relevance pipeline.
      security:
        - LuciaSession: []
      responses:
        '200':
          description: Saved query text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedQueryResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update user's saved search query
      operationId: accountUpdateSavedQuery
      tags:
        - Account
      description: |
        Update the authenticated user's saved search query.
        Re-embeds the query server-side using OpenAI text-embedding-3-small.
      security:
        - LuciaSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSavedQueryRequest'
      responses:
        '200':
          description: Saved query updated and re-embedded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedQueryResponse'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /feedback:
    post:
      summary: Record thumbs feedback (authenticated portal)
      operationId: feedbackRecord
      tags:
        - Feedback
      description: |
        Record 👍/👎 feedback on a DA from the web portal.
        Requires active Lucia session.
        Feedback is used for personalised reranking after ≥200 labelled pairs.
        **FR-024 (Portal thumbs feedback)**: Enable in-portal single-tap feedback.
      security:
        - LuciaSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortalFeedbackRequest'
      responses:
        '200':
          description: Feedback recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /billing/checkout:
    post:
      summary: Create Stripe checkout session
      operationId: billingCheckout
      tags:
        - Billing
      description: |
        Create a Stripe Checkout session for a billing plan (solo or team).
        Returns a Stripe Checkout URL.
        Requires active Lucia session.
        **FR-018 (Subscription checkout)**: Enable Stripe AU billing with GST.
      security:
        - LuciaSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (invalid plan)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /billing/subscription:
    delete:
      summary: Cancel subscription at period end
      operationId: billingCancelSubscription
      tags:
        - Billing
      description: |
        Set `cancel_at_period_end = true` on the user's active Stripe subscription.
        Access continues until the current period ends; no immediate revocation.
        Requires active Lucia session.
        **FR-021 (Subscription cancellation)**: Enable in-app subscription cancellation.
      security:
        - LuciaSession: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 500
                  description: Optional cancellation reason. V1 — logged only, not persisted.
      responses:
        '200':
          description: Subscription set to cancel at period end
          content:
            application/json:
              schema:
                type: object
                required: [ok, accessUntil]
                properties:
                  ok:
                    type: boolean
                    example: true
                  accessUntil:
                    type: string
                    format: date-time
                    description: ISO 8601 timestamp of current_period_end
                    example: '2026-05-28T23:59:59.000Z'
        '401':
          description: Unauthorized (no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active subscription found for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Stripe API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ──────────────────────────────────────────────────────────────────────────────
  # Digests Routes
  # ──────────────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required:
        - error
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        issues:
          type: object
          description: Field-level validation errors
      required:
        - error
        - issues

    # ────────────────────────────────────────────────────────────────────────────
    # Authentication Schemas
    # ────────────────────────────────────────────────────────────────────────────
...
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        lgaBundles:
          type: array
          items:
            type: string
          example: [sydney-inner-west, sydney-northern-beaches]
        savedQueryText:
          type: string
          nullable: true
          example: "Residential roof repairs and reroof work in Sydney"
        smsOptIn:
          type: boolean
          example: true
      required:
        - id
        - userId
    UpdateProfileRequest:
      type: object
      properties:
        mobile_e164:
          type: string
          pattern: '^\+[1-9]\d{6,14}$'
          example: '+61412345678'
      description: Only specified fields are updated
    LgaBundlesResponse:
      type: object
      properties:
        bundle_ids:
          type: array
          items:
            type: string
          example: [sydney-inner-west, sydney-northern-beaches, sydney-western-suburbs]
      required:
        - bundle_ids
    UpdateLgaBundlesRequest:
      type: object
      properties:
        bundle_ids:
          type: array
          items:
            type: string
          minItems: 1
          example: [sydney-inner-west, sydney-northern-beaches]
      required:
        - bundle_ids
    SavedQueryResponse:
      type: object
      properties:
        saved_query_text:
          type: string
          nullable: true
          example: "Residential roof repairs and reroof work in Sydney"
      required:
        - saved_query_text
    UpdateSavedQueryRequest:
      type: object
      properties:
        saved_query_text:
          type: string
          minLength: 1
          maxLength: 500
          example: "Residential roof repairs and reroof work in Sydney"
      required:
        - saved_query_text

    # ────────────────────────────────────────────────────────────────────────────
    # Feedback Schemas
    # ────────────────────────────────────────────────────────────────────────────
    PortalFeedbackRequest:
      type: object
      properties:
        da_id:
          type: string
          example: '2025/123456'
          description: Development Application ID
        feedback:
          type: string
          enum: [up, down, remove]
          example: up
          description: Thumbs up, thumbs down, or remove existing feedback
      required:
        - da_id
        - feedback

    # ────────────────────────────────────────────────────────────────────────────
    # Billing Schemas
    # ────────────────────────────────────────────────────────────────────────────
    CheckoutRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [solo, team]
          example: solo
          description: Solo = AUD 199/mo (1 seat); Team = AUD 499/mo (3 seats)
      required:
        - plan
    CheckoutResponse:
      type: object
      properties:
        checkout_url:
          type: string
          format: uri
          example: 'https://checkout.stripe.com/pay/cs_test_...'
      required:
        - checkout_url
...
  securitySchemes:
    LuciaSession:
      type: apiKey
      in: cookie
      name: lucia_session
      description: Lucia session cookie (set automatically on login/signup)

    CronSecret:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Bearer token for Vercel Cron handlers.
        Value: `Bearer ${CRON_SECRET}` (from GCP Secret Manager).
        Not for client consumption.

    HmacToken:
      type: apiKey
      in: query
      name: token
      description: HMAC-signed token from email feedback links (GET /api/feedback/{token})